feat: Postgres/GCP migration + GCP Identity Platform auth
Squash of feat/postgres-migration. Highlights:

Data/infra
- Single Postgres (Cloud SQL) control plane; goose migrations, sqlc gen.
- Terragrunt/OpenTofu live tree: cloud-sql, cloud-run, artifact-registry,
  secrets, database units across dev/stg/prd on project tallyo-saas.
- Cloud Build (cloudbuild.yaml, BuildKit) builds+pushes the amd64 image.

Auth (GCP Identity Platform)
- Replace password+scs sessions with stateless Firebase bearer-JWT verify.
- Flat user pool: one firebase_uid per person, mapped to many (tenant,role)
  rows; users.password_hash + sessions table dropped (migration 00002).
- New identity-platform tofu module: GCIP config (email/password, email-link,
  Google IdP), web API key; root.hcl gets user_project_override/billing_project.
- RequireAuth middleware (TokenVerifier seam), ResolveTenant by firebase_uid,
  public GET /api/auth/config surfacing firebase config + enabled methods.
- Per-env auth-method flags; dev/stg Cloud Run IAM-locked to owner, prd public.

Frontend (SvelteKit)
- Modular Firebase SDK, runtime config bootstrap, Bearer on every request.
- Roll-own auth UI driven by enabled methods; seamless same-email account
  linking (email/password, email-link, Google) under one uid.

Also drops the recurring-templates slice.

+++ internal/db/migrations/tenant/00003_catalogue.sql
@@ -21,7 +21,7 @@ CREATE TABLE items (
     name                  TEXT NOT NULL,
     unit                  TEXT DEFAULT '',
     category              TEXT,                       -- generic grouping
-    unit_price            REAL,                       -- generic per-unit price (NULL = none/free-form)
+    unit_price            double precision,           -- generic per-unit price (NULL = none/free-form)
     taxable               INTEGER NOT NULL DEFAULT 0, -- ingest always sets this explicitly
     metadata              TEXT DEFAULT '{}',
     UNIQUE (price_list_version_id, code)
